requirements.txt
streamlit
crewai
crewai-tools
tavily-python
langchain-groq
python-dotenv